requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Database ORM
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5

# Data Validation
pydantic==2.4.2

# Scheduling
APScheduler==3.10.4

# SSH/SCP File Transfer
paramiko==3.3.1

# Audio Metadata
mutagen==1.47.0

# Image Processing (for artwork)
Pillow==10.0.1

# Environment Variables
python-dotenv==1.0.0

# Logging
structlog==23.1.0

# HTTP Requests
requests==2.31.0

# Date/Time Utilities
python-dateutil==2.8.2
croniter==1.4.1

# Configuration Management
PyYAML==6.0.1

# System Monitoring
psutil==5.9.5

# Development Dependencies
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0